Moved test files into separate folders, renamed packages so they follow naming convention
