+++ src/main/java/Controllers/MethodOptionsController.java
@@ -1,6 +1,6 @@
-package Controllers;
+package controllers;
 
-import Requests.Request;
+import requests.Request;
 
 public class MethodOptionsController implements Controller{
