+++ src/main/java/Main.java
@@ -1,3 +1,5 @@
+import servers.Server;
+
 import java.io.IOException;
 import java.net.InetAddress;
 import java.net.ServerSocket;
