+++ src/main/java/Requests/Request.java
@@ -1,4 +1,4 @@
-package Requests;
+package requests;
 public class Request {
     private String path;
     private String httpVerb;
