+++ src/tests/controllers/RequestsControllerTest.java
@@ -1,7 +1,8 @@
-import Controllers.Controller;
-import Controllers.LogController;
-import Controllers.TheseController;
-import Requests.Request;
+package controllers;
+
+import controllers.Controller;
+import controllers.RequestsController;
+import requests.Request;
 import httpStatus.HttpStatus;
 import logs.Log;
 import org.junit.Test;
@@ -10,14 +11,15 @@ import specialCharacters.EscapeCharacters;
 import java.io.IOException;
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
+import static junit.framework.TestCase.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-public class TheseControllerTest {
+public class RequestsControllerTest {
 
     @Test
     public void TestRepliesWithTwoHundredOkayForPut() throws IOException {
-        Request request = new Request("/these", "PUT", null, null);
-        Controller controller = new TheseController();
+        Request request = new Request("/requests", "HEAD", null, null);
+        Controller controller = new RequestsController();
         String response = controller.handle(request);
 
         assertEquals(HttpStatus.okay + EscapeCharacters.newline + EscapeCharacters.newline, response);
@@ -26,21 +28,21 @@ public class TheseControllerTest {
     @Test
     public void TestVisitGetsAddedToLog() throws IOException {
         Log log = Log.getInstance();
-        Request request = new Request("/these", "PUT", null, null);
-        Controller controller = new TheseController();
+        Request request = new Request("/requests", "HEAD", null, null);
+        Controller controller = new RequestsController();
 
         controller.handle(request);
 
-        List<String> recentVisits = log.recentVisits(1);
+        List<String> recentVisits = log.recentVisits(100);
         String recentVisit = recentVisits.get(0);
 
-        assertEquals("PUT /these HTTP/1.1", recentVisit);
+        assertEquals("HEAD /requests HTTP/1.1", recentVisit);
     }
 
     @Test
     public void TestMethodsNotAllowed() throws IOException {
-        Request request = new Request("/these", "POST", null, null);
-        Controller controller = new TheseController();
+        Request request = new Request("/requests", "POST", null, null);
+        Controller controller = new RequestsController();
         String response = controller.handle(request);
 
         assertEquals(HttpStatus.methodNotAllowed + EscapeCharacters.newline + EscapeCharacters.newline,
